+++ tencentcloud/iai/v20200303/client.go
@@ -778,6 +778,35 @@ func (c *Client) ModifyPersonGroupInfo(request *ModifyPersonGroupInfoRequest) (r
     return
 }
 
+func NewRevertGroupFaceModelVersionRequest() (request *RevertGroupFaceModelVersionRequest) {
+    request = &RevertGroupFaceModelVersionRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("iai", APIVersion, "RevertGroupFaceModelVersion")
+    return
+}
+
+func NewRevertGroupFaceModelVersionResponse() (response *RevertGroupFaceModelVersionResponse) {
+    response = &RevertGroupFaceModelVersionResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 本接口用于回滚人员库的人脸识别算法模型版本。单个人员库有且仅有一次回滚机会。
+// 
+// 回滚操作会在10s内生效，回滚操作中，您对人员库的操作可能会失效。
+// 
+// 注：给客户我会写10s内生效，我们实际上越快越好。待讨论。
+func (c *Client) RevertGroupFaceModelVersion(request *RevertGroupFaceModelVersionRequest) (response *RevertGroupFaceModelVersionResponse, err error) {
+    if request == nil {
+        request = NewRevertGroupFaceModelVersionRequest()
+    }
+    response = NewRevertGroupFaceModelVersionResponse()
+    err = c.Send(request, response)
+    return
+}
+
 func NewSearchFacesRequest() (request *SearchFacesRequest) {
     request = &SearchFacesRequest{
         BaseRequest: &tchttp.BaseRequest{},