+++ tencentcloud/ocr/v20181119/client.go
@@ -495,7 +495,7 @@ func NewGeneralAccurateOCRResponse() (response *GeneralAccurateOCRResponse) {
 // 
 // 适用于文字较多、版式复杂、对识别准召率要求较高的场景，如试卷试题、网络图片、街景店招牌、法律卷宗等场景。
 // 
-// 产品优势：与通用印刷体识别的基础上，提供更高精度的文字识别服务，在文字较多、长串数字、小字、模糊字、倾斜文本等困难场景下，高精度版的准确率和召回率更高。
+// 产品优势：与通用印刷体识别相比，提供更高精度的文字识别服务，在文字较多、长串数字、小字、模糊字、倾斜文本等困难场景下，高精度版的准确率和召回率更高。
 // 
 // 通用印刷体识别不同版本的差异如下：
 // <table style="width:715px">
@@ -1343,6 +1343,33 @@ func (c *Client) ResidenceBookletOCR(request *ResidenceBookletOCRRequest) (respo
         request = NewResidenceBookletOCRRequest()
     }
     response = NewResidenceBookletOCRResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewRideHailingDriverLicenseOCRRequest() (request *RideHailingDriverLicenseOCRRequest) {
+    request = &RideHailingDriverLicenseOCRRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("ocr", APIVersion, "RideHailingDriverLicenseOCR")
+    return
+}
+
+func NewRideHailingDriverLicenseOCRResponse() (response *RideHailingDriverLicenseOCRResponse) {
+    response = &RideHailingDriverLicenseOCRResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 本接口支持网约车驾驶证重要字段的自动定位与识别，重点字段的识别准确度达到99%以上。
+// 
+// 网约车驾驶证：包括姓名、证号、起始日期、截止日期、发证日期。
+func (c *Client) RideHailingDriverLicenseOCR(request *RideHailingDriverLicenseOCRRequest) (response *RideHailingDriverLicenseOCRResponse, err error) {
+    if request == nil {
+        request = NewRideHailingDriverLicenseOCRRequest()
+    }
+    response = NewRideHailingDriverLicenseOCRResponse()
     err = c.Send(request, response)
     return
 }