+++ tencentcloud/sqlserver/v20180328/models.go
@@ -2438,7 +2438,7 @@ type MigrateTask struct {
 	// 迁移任务的结束时间
 	EndTime *string `json:"EndTime,omitempty" name:"EndTime"`
 
-	// 迁移任务的状态（1:初始化,4:迁移中,5.迁移失败,6.迁移成功）
+	// 迁移任务的状态（1:初始化,4:迁移中,5.迁移失败,6.迁移成功,7已中止,8已删除,9中止中,10完成中,11中止失败,12完成失败）
 	Status *uint64 `json:"Status,omitempty" name:"Status"`
 
 	// 信息